Renaming DetailsContextMenu to ResourceDrawer
DetailsContextMenu is to similar to DetailsMenuContext, and can be confusing (talked with Andy about this). Hopefully this makes it clearer.
